Start working on a safe X11 module (#55)
I had this idea when trying `slice::from_raw_parts` for debugging this part of
the code the other day. If I could encapsulate all of the unsafe code in this
`x` module, then rwm proper could use only safe code. One day this could even be
a separate crate for other window managers to use.